Fix issues

+++ inc/admin/update-post.php
@@ -30,7 +30,9 @@ function scm_update_post( $post_ID, $post_after, $post_before ) {
 		$driver_type = get_option( 'scm_option_driver' );
 		$driver      = scm_driver_factory( $driver_type );
 
-		$driver->delete( $cache_key );
+		if ( $driver ) {
+			$driver->delete( $cache_key );
+		}
 	}
 }
